+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0 - Unreleased
 
+* Remove reopen, close, reprepare, finalize
 * Rename (Database|Statement)Wrapper → *Obj, (Database|Statement).raw → *.ptr, internal sqlite3* types → Sqlite3*
 
 ## 1.3.0 - April 28 2024